Fixed error handling on already done parts
I started doing handleError with the wrong params everywhere :pensive:.... also farts.js has been deprecated :smirk:

+++ modules/Bot/cringe.js
@@ -19,7 +19,7 @@ module.exports = {
       // yeah it just does this
       toggleAliTrust(client, message, args);
       
-      message.delete();
+      if (message.deletable) message.delete();
       //message.react('✅');
     },
 };
